refactor(settings): relocate General tab sections to domain tabs
