+++ web/src/i18n/locales/de.json
@@ -219,7 +219,8 @@
       "calibre": "Calibre",
       "abs": "Audiobookshelf",
       "grimmory": "Grimmory",
-      "blocklist": "Blockliste"
+      "blocklist": "Blockliste",
+      "apiKeys": "API Keys"
     },
     "abs": {
       "review": {
