+++ web/src/i18n/locales/id.json
@@ -238,7 +238,8 @@
       "calibre": "Calibre",
       "abs": "Audiobookshelf",
       "grimmory": "Grimmory",
-      "blocklist": "Daftar blokir"
+      "blocklist": "Daftar blokir",
+      "apiKeys": "API Keys"
     },
     "abs": {
       "review": {
