+++ web/src/i18n/locales/tl.json
@@ -238,7 +238,8 @@
       "calibre": "Calibre",
       "abs": "Audiobookshelf",
       "grimmory": "Grimmory",
-      "blocklist": "Blocklist"
+      "blocklist": "Blocklist",
+      "apiKeys": "API Keys"
     },
     "abs": {
       "review": {
