+++ web/src/pages/settings/GeneralTab.tsx
@@ -1,6 +1,6 @@
 import { FormEvent, useEffect, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { api, BINDERY_BASE, AuthConfig, AuthStatus, AuthorMonitorMode, HardcoverTestResult, SystemStatus } from '../../api/client'
+import { api, AuthConfig, AuthStatus } from '../../api/client'
 import AuthSettings from '../../settings/AuthSettings'
 import ThemeToggle from '../../components/ThemeToggle'
 import LanguageSwitcher from '../../components/LanguageSwitcher'
@@ -8,46 +8,16 @@ import ClipboardManualFallback from '../../components/ClipboardManualFallback'
 import { useClipboardCopy } from '../../components/useClipboardCopy'
 import { useAuth } from '../../auth/AuthContext'
 import { inputCls } from './formStyles'
-import Toggle from './Toggle'
 import NamingTemplateField from './NamingTemplateField'
 import { useSaveResult } from './useSaveResult'
-
-function formatBackupSize(bytes: number): string {
-  if (!bytes || bytes <= 0) return '0 B'
-  const units = ['B', 'KB', 'MB', 'GB', 'TB']
-  const i = Math.min(units.length - 1, Math.floor(Math.log(bytes) / Math.log(1024)))
-  return `${(bytes / Math.pow(1024, i)).toFixed(1)} ${units[i]}`
-}
-
-function formatRelativeTime(iso: string): string {
-  const t = Date.parse(iso)
-  if (isNaN(t)) return ''
-  const diffSec = Math.round((Date.now() - t) / 1000)
-  const abs = Math.abs(diffSec)
-  if (abs < 60) return diffSec >= 0 ? 'just now' : 'in a moment'
-  const mins = Math.round(diffSec / 60)
-  if (Math.abs(mins) < 60) return mins >= 0 ? `${mins}m ago` : `in ${-mins}m`
-  const hrs = Math.round(diffSec / 3600)
-  if (Math.abs(hrs) < 24) return hrs >= 0 ? `${hrs}h ago` : `in ${-hrs}h`
-  const days = Math.round(diffSec / 86400)
-  return days >= 0 ? `${days}d ago` : `in ${-days}d`
-}
-
-function isAuthorMonitorMode(value: string): value is AuthorMonitorMode {
-  return value === 'all' || value === 'future' || value === 'latest' || value === 'none'
-}
 
 export default function GeneralTab() {
   const { t } = useTranslation()
   const { isAdmin } = useAuth()
-  const opdsClipboard = useClipboardCopy()
   const [settings, setSettings] = useState<Record<string, string>>({})
   const [loading, setLoading] = useState(true)
   const [saving, setSaving] = useState<string | null>(null)
   const [dropErr, setDropErr] = useState<string | null>(null)
-  const [backups, setBackups] = useState<Array<{ name: string; size: number; modTime: string }>>([])
-  const [creatingBackup, setCreatingBackup] = useState(false)
-  const [deletingBackup, setDeletingBackup] = useState<string | null>(null)
   const [scanningLibrary, setScanningLibrary] = useState(false)
   const [scanMessage, setScanMessage] = useState<string | null>(null)
   const scanStartedAt = useRef<number>(0)
@@ -65,11 +35,7 @@ export default function GeneralTab() {
     scan_error?: string
   } | null>(null)
   const [storage, setStorage] = useState<{ downloadDir: string; audiobookDownloadDir: string; libraryDir: string; audiobookDir: string } | null>(null)
-  const [systemStatus, setSystemStatus] = useState<SystemStatus | null>(null)
-  const [hardcoverToken, setHardcoverToken] = useState('')
-  const [hardcoverTestResult, setHardcoverTestResult] = useState<(HardcoverTestResult & { testing?: boolean }) | null>(null)
   const [langSaveResult, langSave] = useSaveResult()
-  const [logRetentionSaveResult, logRetentionSave] = useSaveResult()
   const [dropSaveResult, dropSave] = useSaveResult()
 
   useEffect(() => {
@@ -81,19 +47,9 @@ export default function GeneralTab() {
       })
       .catch(console.error)
       .finally(() => setLoading(false))
-    api.listBackups().then(setBackups).catch(console.error)
     api.libraryScanStatus().then(setLastScan).catch(() => {/* no prior scan — ignore 404 */})
     api.getStorage().then(setStorage).catch(console.error)
-    api.status().then(setSystemStatus).catch(console.error)
   }, [])
-
-  const refreshSystemStatus = async () => {
-    try {
-      setSystemStatus(await api.status())
-    } catch (err) {
-      console.error(err)
-    }
-  }
 
   const saveSetting = async (key: string) => {
     setSaving(key)
@@ -120,97 +76,6 @@ export default function GeneralTab() {
       throw err
     } finally {
       setSaving(null)
-    }
-  }
-
-  const saveHardcoverToken = async () => {
-    setSaving('hardcover.api_token')
-    setHardcoverTestResult(null)
-    try {
-      await api.setSetting('hardcover.api_token', hardcoverToken.trim())
-      setHardcoverToken('')
-      await refreshSystemStatus()
-    } catch (err) {
-      console.error(err)
-    } finally {
-      setSaving(null)
-    }
-  }
-
-  const clearHardcoverToken = async () => {
-    setSaving('hardcover.api_token')
-    setHardcoverTestResult(null)
-    try {
-      await api.setSetting('hardcover.api_token', '')
-      setHardcoverToken('')
-      await refreshSystemStatus()
-    } catch (err) {
-      console.error(err)
-    } finally {
-      setSaving(null)
-    }
-  }
-
-  const testHardcover = async () => {
-    setHardcoverTestResult({
-      ok: false,
-      tokenConfigured: hardcoverTokenConfigured,
-      searchResults: 0,
-      catalogOk: false,
-      testing: true,
-    })
-    try {
-      const result = await api.testHardcover()
-      setHardcoverTestResult(result)
-    } catch (err) {
-      setHardcoverTestResult({
-        ok: false,
-        tokenConfigured: hardcoverTokenConfigured,
-        searchResults: 0,
-        catalogOk: false,
-        error: err instanceof Error ? err.message : 'Hardcover API test failed',
-      })
-    }
-  }
-
-  const toggleEnhancedHardcover = async () => {
-    const current = (settings['hardcover.enhanced_series_enabled'] ?? 'false').toLowerCase()
-    const next = current === 'true' ? 'false' : 'true'
-    setSettings(s => ({ ...s, 'hardcover.enhanced_series_enabled': next }))
-    setSaving('hardcover.enhanced_series_enabled')
-    try {
-      await api.setSetting('hardcover.enhanced_series_enabled', next)
-      await refreshSystemStatus()
-    } catch (err) {
-      console.error(err)
-    } finally {
-      setSaving(null)
-    }
-  }
-
-  const handleBackup = async () => {
-    setCreatingBackup(true)
-    try {
-      const result = await api.createBackup()
-      setBackups(prev => [result, ...prev])
-      alert(`Backup created: ${result.name}`)
-    } catch (err) {
-      alert('Backup failed: ' + (err instanceof Error ? err.message : 'Unknown error'))
-    } finally {
-      setCreatingBackup(false)
-    }
-  }
-
-  const handleDeleteBackup = async (filename: string) => {
-    if (!confirm(`Delete backup ${filename}?`)) return
-    setDeletingBackup(filename)
-    try {
-      await api.deleteBackup(filename)
-      setBackups(prev => prev.filter(b => b.name !== filename))
-    } catch (err) {
-      alert('Delete failed: ' + (err instanceof Error ? err.message : 'Unknown error'))
-    } finally {
-      setDeletingBackup(null)
     }
   }
 
@@ -253,18 +118,6 @@ export default function GeneralTab() {
   }
 
   if (loading) return <div className="text-slate-600 dark:text-zinc-500">{t('common.loading')}</div>
-
-  const enhancedHardcoverAdminEnabled = (settings['hardcover.enhanced_series_enabled'] ?? 'false').toLowerCase() === 'true'
-  const hardcoverTokenConfigured = systemStatus?.hardcoverTokenConfigured ?? false
-  const enhancedHardcoverEnabled = systemStatus?.enhancedHardcoverApi ?? false
-  const enhancedHardcoverReason = systemStatus?.enhancedHardcoverDisabledReason
-  const enhancedHardcoverStatus = enhancedHardcoverEnabled
-    ? t('settings.general.enhancedHardcoverStatusEnabled', 'Enabled')
-    : enhancedHardcoverReason === 'env_disabled'
-      ? t('settings.general.enhancedHardcoverStatusEnvDisabled', 'Disabled by BINDERY_ENHANCED_HARDCOVER_API=false.')
-      : enhancedHardcoverReason === 'missing_token'
-        ? t('settings.general.enhancedHardcoverStatusMissingToken', 'Add a Hardcover API token to enable this feature.')
-        : t('settings.general.enhancedHardcoverStatusAdminDisabled', 'Turn this on to enable enhanced series features.')
 
   return (
     <div className="space-y-8">
@@ -624,385 +477,6 @@ export default function GeneralTab() {
         </div>
       </section>
 
-      {/* Metadata provider */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.metadataProvider', 'Metadata Provider')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900">
-          <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200 mb-1">
-            {t('settings.general.metadataProviderLabel', 'Primary metadata provider')}
-          </label>
-          <p className="text-xs text-slate-600 dark:text-zinc-500 mb-2">
-            {t('settings.general.metadataProviderHint', 'Selects the source used for author and book search. DNB (Deutsche Nationalbibliothek) is recommended for German, Austrian, and Swiss catalogues — it covers German-language publications since 1913 where OpenLibrary coverage is thin. OpenLibrary remains the default for other regions. The non-primary provider is always used as an enricher.')}
-          </p>
-          <select
-            value={settings['metadata.primary_provider'] ?? 'openlibrary'}
-            onChange={async e => {
-              const next = e.target.value
-              setSettings(s => ({ ...s, 'metadata.primary_provider': next }))
-              await api.setSetting('metadata.primary_provider', next).catch(console.error)
-            }}
-            className="bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded-md px-3 py-2 text-sm focus:outline-none focus:border-emerald-500"
-          >
-            <option value="openlibrary">{t('settings.general.metadataProviderOpenlibrary', 'OpenLibrary (default)')}</option>
-            <option value="dnb">{t('settings.general.metadataProviderDnb', 'DNB — Deutsche Nationalbibliothek (German/DACH)')}</option>
-          </select>
-        </div>
-      </section>
-
-      {/* Author defaults */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.authorDefaults', 'Author defaults')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900 space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200 mb-1">
-              {t('settings.general.defaultMediaTypeLabel', 'Default media type')}
-            </label>
-            <p className="text-xs text-slate-600 dark:text-zinc-500 mb-2">
-              {t('settings.general.defaultMediaTypeHint', 'Applied to new authors when no explicit choice is made. Existing authors are unaffected — use the Authors page bulk action to migrate them.')}
-            </p>
-            <select
-              value={settings['default.media_type'] ?? 'ebook'}
-              onChange={async e => {
-                const next = e.target.value
-                setSettings(s => ({ ...s, 'default.media_type': next }))
-                await api.setSetting('default.media_type', next).catch(console.error)
-              }}
-              className="bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded-md px-3 py-2 text-sm focus:outline-none focus:border-emerald-500"
-            >
-              <option value="ebook">{t('mediaType.ebook', 'Ebook')}</option>
-              <option value="audiobook">{t('mediaType.audiobook', 'Audiobook')}</option>
-              <option value="both">{t('mediaType.both', 'Both')}</option>
-            </select>
-          </div>
-          <div>
-            <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200 mb-1">
-              {t('settings.general.defaultMonitorModeLabel', 'Default monitor mode')}
-            </label>
-            <p className="text-xs text-slate-600 dark:text-zinc-500 mb-2">
-              {t('settings.general.defaultMonitorModeHint', 'Applied to newly added authors. Existing authors keep their current mode unless edited.')}
-            </p>
-            <select
-              value={isAuthorMonitorMode(settings['author.default_monitor_mode'] ?? '') ? settings['author.default_monitor_mode'] : 'all'}
-              onChange={async e => {
-                const next = e.target.value as AuthorMonitorMode
-                setSettings(s => ({ ...s, 'author.default_monitor_mode': next }))
-                await api.setSetting('author.default_monitor_mode', next).catch(console.error)
-              }}
-              className="bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded-md px-3 py-2 text-sm focus:outline-none focus:border-emerald-500"
-            >
-              <option value="all">{t('monitorMode.all', 'All books')}</option>
-              <option value="future">{t('monitorMode.future', 'Future books only')}</option>
-              <option value="latest">{t('monitorMode.latest', 'Latest only')}</option>
-              <option value="none">{t('monitorMode.none', 'None')}</option>
-            </select>
-          </div>
-          {(settings['author.default_monitor_mode'] ?? 'all') === 'latest' && (
-            <div>
-              <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200 mb-1">
-                {t('settings.general.defaultMonitorLatestCountLabel', 'Latest book count')}
-              </label>
-              <input
-                type="number"
-                min={1}
-                value={settings['author.default_monitor_latest_count'] ?? '1'}
-                onChange={async e => {
-                  const next = e.target.value
-                  setSettings(s => ({ ...s, 'author.default_monitor_latest_count': next }))
-                  if (/^[1-9]\d*$/.test(next)) {
-                    await api.setSetting('author.default_monitor_latest_count', next).catch(console.error)
-                  }
-                }}
-                className="w-28 bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded-md px-3 py-2 text-sm focus:outline-none focus:border-emerald-500"
-              />
-            </div>
-          )}
-        </div>
-      </section>
-
-      {/* Auto-grab */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.autoGrab')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900">
-          <div className="flex items-center justify-between">
-            <div>
-              <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200">{t('settings.general.autoGrabLabel')}</label>
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">{t('settings.general.autoGrabHint')}</p>
-            </div>
-            <Toggle
-              checked={(settings['autoGrab.enabled'] ?? 'true') !== 'false'}
-              onChange={async () => {
-                const current = (settings['autoGrab.enabled'] ?? 'true').toLowerCase()
-                const next = current === 'false' ? 'true' : 'false'
-                setSettings(s => ({ ...s, 'autoGrab.enabled': next }))
-                await api.setSetting('autoGrab.enabled', next).catch(console.error)
-              }}
-              title={(settings['autoGrab.enabled'] ?? 'true') !== 'false' ? t('common.disable') : t('common.enable')}
-            />
-          </div>
-        </div>
-      </section>
-
-      {/* Recommendations */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.recommendations')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900">
-          <div className="flex items-center justify-between">
-            <div>
-              <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200">{t('settings.general.recommendationsLabel')}</label>
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">{t('settings.general.recommendationsHint')}</p>
-            </div>
-            <Toggle
-              checked={(settings['recommendations.enabled'] ?? 'false') === 'true'}
-              onChange={async () => {
-                const current = (settings['recommendations.enabled'] ?? 'false').toLowerCase()
-                const next = current === 'true' ? 'false' : 'true'
-                setSettings(s => ({ ...s, 'recommendations.enabled': next }))
-                await api.setSetting('recommendations.enabled', next).catch(console.error)
-              }}
-              title={(settings['recommendations.enabled'] ?? 'false') === 'true' ? t('common.disable') : t('common.enable')}
-            />
-          </div>
-        </div>
-      </section>
-
-      {/* OIDC providers */}
-      <AuthSettings />
-
-      {/* API Keys */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.apiKeys')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900 space-y-4">
-          <div>
-            <label className="block text-xs text-slate-600 dark:text-zinc-400 mb-1">{t('settings.general.googleBooksKey')}</label>
-            <div className="flex gap-2">
-              <input
-                value={settings['googlebooks.apiKey'] ?? ''}
-                onChange={e => setSettings(s => ({ ...s, 'googlebooks.apiKey': e.target.value }))}
-                placeholder="AIza..."
-                type="password"
-                className="flex-1 bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded px-3 py-2 text-sm focus:outline-none focus:border-slate-400 dark:focus:border-zinc-600"
-              />
-              <button
-                onClick={() => saveSetting('googlebooks.apiKey')}
-                disabled={saving === 'googlebooks.apiKey'}
-                className="px-3 py-2 bg-emerald-600 hover:bg-emerald-500 rounded text-xs font-medium disabled:opacity-50"
-              >
-                {saving === 'googlebooks.apiKey' ? t('common.saving') : t('common.save')}
-              </button>
-            </div>
-          </div>
-
-          <div className="border-t border-slate-200 dark:border-zinc-800 pt-4 space-y-3">
-            <div>
-              <div className="flex items-center justify-between gap-3 mb-1">
-                <label className="block text-xs text-slate-600 dark:text-zinc-400">
-                  {t('settings.general.hardcoverApiToken', 'Hardcover API Token')}
-                </label>
-                <span className={`text-[11px] font-medium ${hardcoverTokenConfigured ? 'text-emerald-600 dark:text-emerald-400' : 'text-slate-500 dark:text-zinc-500'}`}>
-                  {hardcoverTokenConfigured
-                    ? t('settings.general.hardcoverTokenConfigured', 'Token configured')
-                    : t('settings.general.hardcoverTokenNotConfigured', 'No token configured')}
-                </span>
-              </div>
-              <div className="flex gap-2">
-                <input
-                  value={hardcoverToken}
-                  onChange={e => setHardcoverToken(e.target.value)}
-                  placeholder={hardcoverTokenConfigured
-                    ? t('settings.general.hardcoverApiTokenConfiguredPlaceholder', 'Saved token is hidden. Enter a new token to rotate it.')
-                    : t('settings.general.hardcoverApiTokenPlaceholder', 'Paste a Hardcover API token')}
-                  type="password"
-                  className="flex-1 bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded px-3 py-2 text-sm focus:outline-none focus:border-slate-400 dark:focus:border-zinc-600"
-                />
-                <button
-                  onClick={saveHardcoverToken}
-                  disabled={saving === 'hardcover.api_token' || !hardcoverToken.trim()}
-                  className="px-3 py-2 bg-emerald-600 hover:bg-emerald-500 rounded text-xs font-medium disabled:opacity-50"
-                  aria-label={t('settings.general.hardcoverSaveToken', 'Save Hardcover API token')}
-                >
-                  {saving === 'hardcover.api_token' ? t('common.saving') : t('common.save')}
-                </button>
-                {hardcoverTokenConfigured && (
-                  <button
-                    onClick={clearHardcoverToken}
-                    disabled={saving === 'hardcover.api_token'}
-                    className="px-3 py-2 bg-slate-300 dark:bg-zinc-700 hover:bg-slate-400 dark:hover:bg-zinc-600 rounded text-xs font-medium disabled:opacity-50"
-                    aria-label={t('settings.general.hardcoverClearToken', 'Clear Hardcover API token')}
-                  >
-                    {t('settings.general.hardcoverClearToken', 'Clear')}
-                  </button>
-                )}
-                <button
-                  onClick={testHardcover}
-                  disabled={!hardcoverTokenConfigured || hardcoverTestResult?.testing || saving === 'hardcover.api_token'}
-                  className="px-3 py-2 bg-slate-300 dark:bg-zinc-700 hover:bg-slate-400 dark:hover:bg-zinc-600 rounded text-xs font-medium disabled:opacity-50"
-                  aria-label={t('settings.general.hardcoverTestApi', 'Test Hardcover API')}
-                >
-                  {hardcoverTestResult?.testing ? t('common.testing', 'Testing...') : t('common.test', 'Test')}
-                </button>
-              </div>
-              <a
-                href="https://hardcover.app/account/api"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="inline-block mt-1 text-xs text-emerald-600 dark:text-emerald-400 hover:underline"
-              >
-                {t('settings.general.hardcoverApiTokenLink', 'Create or copy a Hardcover API token')}
-              </a>
-              {hardcoverTestResult && !hardcoverTestResult.testing && (
-                <p className={`mt-1 text-xs ${hardcoverTestResult.ok ? 'text-emerald-600 dark:text-emerald-400' : 'text-rose-600 dark:text-rose-400'}`}>
-                  {hardcoverTestResult.ok
-                    ? (hardcoverTestResult.message || t('settings.general.hardcoverTestOk', 'Hardcover API connected.'))
-                    : (hardcoverTestResult.error || t('settings.general.hardcoverTestFailed', 'Hardcover API test failed.'))}
-                </p>
-              )}
-            </div>
-
-            <div className="flex items-center justify-between gap-4">
-              <div className="min-w-0">
-                <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200">
-                  {t('settings.general.enhancedHardcoverSeries', 'Enhanced Hardcover series')}
-                </label>
-                <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">
-                  {t('settings.general.enhancedHardcoverSeriesHint', 'Use Hardcover-backed series matching, catalog diffs, and missing-book fill.')}
-                </p>
-                <p className={`text-xs mt-1 ${enhancedHardcoverEnabled ? 'text-emerald-600 dark:text-emerald-400' : 'text-slate-600 dark:text-zinc-500'}`}>
-                  {enhancedHardcoverStatus}
-                </p>
-              </div>
-              <Toggle
-                checked={enhancedHardcoverAdminEnabled}
-                onChange={toggleEnhancedHardcover}
-                disabled={saving === 'hardcover.enhanced_series_enabled'}
-                title={enhancedHardcoverAdminEnabled ? t('common.disable') : t('common.enable')}
-              />
-            </div>
-          </div>
-        </div>
-      </section>
-
-      {/* OPDS */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.opds')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900">
-          <div>
-            <span className="block text-xs font-medium text-slate-600 dark:text-zinc-400 mb-1">{t('settings.general.opdsFeedUrl')}</span>
-            <div className="flex items-center gap-2">
-              <code className="flex-1 text-xs bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 px-2 py-1.5 rounded font-mono break-all">
-                {window.location.origin}{BINDERY_BASE}/opds
-              </code>
-              <button
-                onClick={() => opdsClipboard.copy(window.location.origin + BINDERY_BASE + '/opds')}
-                className="px-3 py-1.5 bg-slate-600 hover:bg-slate-500 rounded text-xs font-medium flex-shrink-0"
-              >
-                {opdsClipboard.status === 'copied' ? t('common.copied', 'Copied') : t('settings.general.copy')}
-              </button>
-            </div>
-            {opdsClipboard.status === 'manual' && (
-              <ClipboardManualFallback text={opdsClipboard.manualText} className="mt-2" />
-            )}
-            <p className="text-xs text-slate-500 dark:text-zinc-500 mt-1.5">{t('settings.general.opdsHint')}</p>
-          </div>
-        </div>
-      </section>
-
-      {/* Log retention */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.logRetention')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900">
-          <div className="flex items-center justify-between gap-4">
-            <div>
-              <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200">{t('settings.general.logRetentionLabel')}</label>
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">{t('settings.general.logRetentionHint')}</p>
-            </div>
-            <div className="flex items-center gap-2 flex-shrink-0">
-              <input
-                type="number"
-                min={1}
-                max={365}
-                value={settings['log.retention_days'] ?? '14'}
-                onChange={e => setSettings(s => ({ ...s, 'log.retention_days': e.target.value }))}
-                className="w-20 bg-slate-200 dark:bg-zinc-800 border border-slate-300 dark:border-zinc-700 rounded px-2 py-1 text-sm text-right"
-              />
-              <span className="text-sm text-slate-600 dark:text-zinc-400">{t('settings.general.days')}</span>
-              <button
-                onClick={() => logRetentionSave(() => api.setSetting('log.retention_days', settings['log.retention_days'] ?? ''))}
-                disabled={saving === 'log.retention_days'}
-                className={`px-3 py-1.5 rounded text-sm font-medium disabled:opacity-50 ${logRetentionSaveResult === 'saved' ? 'bg-emerald-500' : logRetentionSaveResult === 'error' ? 'bg-red-600' : 'bg-emerald-600 hover:bg-emerald-500'}`}
-              >
-                {logRetentionSaveResult === 'saved' ? 'Saved ✓' : logRetentionSaveResult === 'error' ? 'Error' : saving === 'log.retention_days' ? t('common.saving') : t('common.save')}
-              </button>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      {/* Backup */}
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.backup')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900 space-y-3">
-          <div className="flex items-center justify-between">
-            <div>
-              <p className="text-sm text-slate-700 dark:text-zinc-300">{t('settings.general.backupCreate')}</p>
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">{t('settings.general.backupHint')}</p>
-            </div>
-            <button
-              onClick={handleBackup}
-              disabled={creatingBackup}
-              className="px-4 py-2 bg-emerald-600 hover:bg-emerald-500 rounded text-sm font-medium disabled:opacity-50 flex-shrink-0"
-            >
-              {creatingBackup ? t('settings.general.backupCreating') : t('settings.general.backupButton')}
-            </button>
-          </div>
-          {backups.length > 0 && (
-            <div className="mt-3 border-t border-slate-200 dark:border-zinc-800 pt-3">
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mb-2">{t('settings.general.existingBackups')}</p>
-              <ul className="space-y-1">
-                {backups.map(b => (
-                  <li key={b.name} className="flex items-center justify-between text-xs text-slate-600 dark:text-zinc-400">
-                    <span>
-                      <span className="font-mono">{b.name}</span>
-                      <span className="ml-2 text-slate-500 dark:text-zinc-500">{formatBackupSize(b.size)} · {formatRelativeTime(b.modTime)}</span>
-                    </span>
-                    <button
-                      onClick={() => handleDeleteBackup(b.name)}
-                      disabled={deletingBackup === b.name}
-                      className="ml-4 text-red-600 dark:text-red-400 hover:underline disabled:opacity-50"
-                    >
-                      {deletingBackup === b.name ? '…' : t('common.delete')}
-                    </button>
-                  </li>
-                ))}
-              </ul>
-            </div>
-          )}
-        </div>
-      </section>
-
-      <section>
-        <h3 className="text-base font-semibold mb-3 text-slate-800 dark:text-zinc-200">{t('settings.general.telemetry')}</h3>
-        <div className="p-4 border border-slate-200 dark:border-zinc-800 rounded-lg bg-slate-100 dark:bg-zinc-900 space-y-3">
-          <div className="flex items-start justify-between gap-4">
-            <div>
-              <label className="block text-sm font-medium text-slate-800 dark:text-zinc-200">{t('settings.general.telemetryLabel')}</label>
-              <p className="text-xs text-slate-600 dark:text-zinc-500 mt-0.5">{t('settings.general.telemetryHint')}</p>
-            </div>
-            <Toggle
-              checked={(settings['telemetry.enabled'] ?? 'true') !== 'false'}
-              onChange={async () => {
-                const current = (settings['telemetry.enabled'] ?? 'true').toLowerCase()
-                const next = current !== 'false' ? 'false' : 'true'
-                setSettings(s => ({ ...s, 'telemetry.enabled': next }))
-                await api.setSetting('telemetry.enabled', next).catch(console.error)
-              }}
-              title={(settings['telemetry.enabled'] ?? 'true') !== 'false' ? t('common.disable') : t('common.enable')}
-            />
-          </div>
-          <p className="text-xs text-slate-500 dark:text-zinc-600">
-            {t('settings.general.telemetryDetail')}
-          </p>
-        </div>
-      </section>
       </>)}
     </div>
   )
@@ -1136,6 +610,12 @@ function SecuritySection() {
         {status?.authenticated && (
           <div className={isAdmin ? 'border-t border-slate-200 dark:border-zinc-800 pt-4' : ''}>
             <ChangePasswordForm username={cfg.username} />
+          </div>
+        )}
+
+        {isAdmin && (
+          <div className="border-t border-slate-200 dark:border-zinc-800 pt-4">
+            <AuthSettings />
           </div>
         )}
       </div>
